+++ packages/awino-ui/src/components/layout/Drawer/Drawer.tsx
@@ -57,6 +57,7 @@ const StyledSwipeableDrawer = styled(SwipeableDrawer)(({ theme }) => ({
     '.right': {
       display: 'flex',
       flexDirection: 'row',
+      alignItems: 'center',
     },
   },
 }));